telegram-channel-monitor/requirements.txt
telethon==1.36.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
fastapi==0.104.1
uvicorn==0.24.0
